K8SPXC-725&763: delete resources consistently using UID

+++ pkg/controller/pxc/controller.go
@@ -1032,7 +1032,7 @@ func (r *ReconcilePerconaXtraDBCluster) deleteServices(svcs ...*corev1.Service) 
 			continue
 		}
 
-		err = r.client.Delete(context.TODO(), s)
+		err = r.client.Delete(context.TODO(), s, &client.DeleteOptions{Preconditions: &metav1.Preconditions{UID: &s.UID}})
 		if err != nil {
 			return errors.Wrapf(err, "delete service: %s", s.Name)
 		}
@@ -1054,7 +1054,7 @@ func (r *ReconcilePerconaXtraDBCluster) deletePVC(namespace string, lbls map[str
 	}
 
 	for _, pvc := range list.Items {
-		err := r.client.Delete(context.TODO(), &pvc)
+		err := r.client.Delete(context.TODO(), &pvc, &client.DeleteOptions{Preconditions: &metav1.Preconditions{UID: &pvc.UID}})
 		if err != nil {
 			return errors.Wrapf(err, "delete PVC %s", pvc.Name)
 		}
@@ -1081,7 +1081,7 @@ func (r *ReconcilePerconaXtraDBCluster) deleteSecrets(cr *api.PerconaXtraDBClust
 			continue
 		}
 
-		err = r.client.Delete(context.TODO(), secret)
+		err = r.client.Delete(context.TODO(), secret, &client.DeleteOptions{Preconditions: &metav1.Preconditions{UID: &secret.UID}})
 		if err != nil {
 			return errors.Wrapf(err, "delete secret %s", secretName)
 		}
